Improved the lockpicking system, it is now randomized and doesn't rely on pre-made textures.
- Lockpicking timing and forgiveness is randomized, there is a randomized delay between each stage of the animation progress and certain amount of time allowed to be late before pressing the right key
- Changed the design, core design is still the same but the key images are on top instead of beside the circle
(see: https://i.imgur.com/UfAOGR4.png)

- Removed debug logging

+++ core/src/main/java/me/vrekt/oasis/GameManager.java
@@ -148,7 +148,7 @@ public class GameManager {
      * @param action the action
      * @param delay  the delay
      */
-    public static void executeTaskLater(Runnable action, long delay) {
+    public static void executeTaskLater(Runnable action, float delay) {
         Timer.schedule(new Timer.Task() {
             @Override
             public void run() {
